Move the commands to a new file in the API folder.
